+++ packages/envex/src/nextjs/EnvScript/index.tsx
@@ -1,6 +1,6 @@
 import Script from 'next/script'
 import { getPublicEnv } from '../utils'
-import type { EnvScriptProps } from './types.ts'
+import type { EnvScriptProps } from './types'
 
 const EnvScript = async ({ scan, schema }: EnvScriptProps = {}) => {
   const env = await getPublicEnv({ scan, schema })
