+++ packages/envex/src/nextjs/EnvScript/types.ts
@@ -1,8 +1,1 @@
-import type { StandardSchemaV1 } from '@standard-schema/spec'
-import type { ScanConfig } from '../../types.ts'
-
-export interface EnvScriptProps {
-  /** Enable credential scanning: `true` for defaults or a config object. Off by default. */
-  scan?: ScanConfig
-  schema?: StandardSchemaV1
-}
+export type { EnvScriptProps } from '../types'
