+++ packages/envex/src/nextjs/InlineEnvScript/index.tsx
@@ -1,5 +1,5 @@
 import { getPublicEnv } from '../utils'
-import type { InlineEnvScriptProps } from './types.ts'
+import type { InlineEnvScriptProps } from './types'
 
 const InlineEnvScript = async ({ scan, schema }: InlineEnvScriptProps = {}) => {
   const env = await getPublicEnv({ scan, schema })
