+++ packages/envex/src/nextjs/InlineEnvScript/types.ts
@@ -1,8 +1,1 @@
-import type { StandardSchemaV1 } from '@standard-schema/spec'
-import type { ScanConfig } from '../../types.ts'
-
-export interface InlineEnvScriptProps {
-  /** Enable credential scanning: `true` for defaults or a config object. Off by default. */
-  scan?: ScanConfig
-  schema?: StandardSchemaV1
-}
+export type { EnvScriptProps as InlineEnvScriptProps } from '../types'
