+++ packages/envex/src/nextjs/utils/getEnv/index.ts
@@ -2,12 +2,7 @@ import type { StandardSchemaV1 } from '@standard-schema/spec'
 import { connection } from 'next/server'
 import type { Env } from '../../../types.ts'
 import { validateEnv } from '../../../utils'
-
-export interface GetEnvOptions<
-  TSchema extends StandardSchemaV1 | undefined = undefined,
-> {
-  schema?: TSchema
-}
+import type { GetEnvOptions } from './types.ts'
 
 const getEnv = async <TSchema extends StandardSchemaV1 | undefined = undefined>(
   options?: GetEnvOptions<TSchema>
